# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API of Alexandria.",
        "title": "Alexandria API",
        "contact": {},
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v2",
    "paths": {
...
        "/project-posts": {
            "put": {
                "description": "Update any number of the aspects of a project post",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "project-posts"
                ],
                "summary": "Update project post",
                "parameters": [
                    {
                        "description": "Updated Project Post",
                        "name": "post",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ProjectPostDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new project post",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "project-posts"
                ],
                "summary": "Create new project post",
                "parameters": [
                    {
                        "description": "Project Post Creation Form",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/forms.ProjectPostCreationForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ProjectPostDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    }
                }
            }
        },
        "/project-posts/from-github": {
            "post": {
                "description": "Create a new project post\nCreates a project post in the same way as CreateProjectPost\nHowever, the post files are imported from the given Github repository",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "project-posts"
                ],
                "summary": "Create new project post with the version imported from github",
                "parameters": [
                    {
                        "description": "Post Creation Form",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/forms.ProjectPostCreationForm"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Github repository url",
                        "name": "url",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ProjectPostDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    }
                }
            }
        },
        "/project-posts/{postID}": {
            "get": {
                "description": "Get a project post by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "project-posts"
                ],
                "summary": "Get project post",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Post ID",
                        "name": "postID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ProjectPostDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a project post with given ID from database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "project-posts"
                ],
                "summary": "Delete a project post",
                "parameters": [
                    {
                        "type": "string",
                        "description": "post ID",
                        "name": "postID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    }
                }
            }
        },
        "/project-posts/{postID}/all-discussions": {
            "get": {
                "description": "Returns all discussion IDs on this project post over all its previous versions, instead of only the current version",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "project-posts"
                ],
                "summary": "Returns all discussion IDs associated with the project post",
                "parameters": [
                    {
                        "type": "string",
                        "description": "post ID",
                        "name": "postID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    }
                }
            }
        },
        "/project-posts/{postID}/merge-requests-by-status": {
            "get": {
                "description": "Returns all MR IDs of this project post, grouped by each MR's review status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "project-posts"
                ],
                "summary": "Returns MR IDs grouped by each MR status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "post ID",
                        "name": "postID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/forms.GroupedMergeRequestForm"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "forms.GroupedMergeRequestForm": {
            "type": "object",
            "properties": {
                "openForReviewIDs": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "peerReviewedIDs": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "rejectedIDs": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "forms.PostCreationForm": {
            "type": "object",
            "properties": {
                "anonymous": {
                    "type": "boolean"
                },
                "authorMemberIDs": {
                    "description": "Members that are authors of the post",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "postType": {
                    "$ref": "#/definitions/tags.PostType"
                },
                "scientificFieldTags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/tags.ScientificField"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "forms.ProjectPostCreationForm": {
            "type": "object",
            "properties": {
                "completionStatus": {
                    "$ref": "#/definitions/tags.CompletionStatus"
                },
                "feedbackPreference": {
                    "$ref": "#/definitions/tags.FeedbackPreference"
                },
                "postCreationForm": {
                    "$ref": "#/definitions/forms.PostCreationForm"
                }
            }
        },
...
        "models.PostDTO": {
            "type": "object",
            "properties": {
                "anonymous": {
                    "type": "boolean"
                },
                "collaboratorIDs": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "postType": {
                    "$ref": "#/definitions/tags.PostType"
                },
                "scientificFieldTags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/tags.ScientificField"
                    }
                },
                "title": {
                    "type": "string"
                },
                "versionID": {
                    "type": "integer"
                }
            }
        },
        "models.ProjectPostDTO": {
            "type": "object",
            "properties": {
                "closedMergeRequestIDs": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "completionStatus": {
                    "$ref": "#/definitions/tags.CompletionStatus"
                },
                "feedbackPreference": {
                    "$ref": "#/definitions/tags.FeedbackPreference"
                },
                "id": {
                    "type": "integer"
                },
                "openMergeRequestIDs": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "postDTO": {
                    "$ref": "#/definitions/models.PostDTO"
                },
                "postReviewStatusTag": {
                    "$ref": "#/definitions/tags.PostReviewStatus"
                }
            }
        },
...
        "tags.CompletionStatus": {
            "type": "string",
            "enum": [
                "idea",
                "ongoing",
                "completed"
            ],
            "x-enum-varnames": [
                "Idea",
                "Ongoing",
                "Completed"
            ]
        },
        "tags.FeedbackPreference": {
            "type": "string",
            "enum": [
                "discussion",
                "formal feedback"
            ],
            "x-enum-varnames": [
                "Discussion",
                "FormalFeedback"
            ]
        },
        "tags.PostReviewStatus": {
            "type": "string",
            "enum": [
                "open",
                "revision needed",
                "reviewed"
            ],
            "x-enum-varnames": [
                "Open",
                "RevisionNeeded",
                "Reviewed"
            ]
        },
        "tags.PostType": {
            "type": "string",
            "enum": [
                "project",
                "question",
                "reflection"
            ],
            "x-enum-varnames": [
                "Project",
                "Question",
                "Reflection"
            ]
        },
        "tags.ScientificField": {
            "type": "string",
            "enum": [
                "mathematics",
                "computer science"
            ],
            "x-enum-varnames": [
                "Mathematics",
                "ComputerScience"
            ]
        },
...
        "utils.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}